* Window finding, showing, hiding, activating, etc. now works. * Stack traces are now logged for errors in callbacks. * Run command exists. (for running arbitrary shell commands) * Shortcut added that runs app in Admin mode. (needed for certain functions, e.g. window manipulations and reliable keyboard-hooking... took me a LONG time to figure out [since I thought I'd tried it as admin earlier to no effect; probably there was another smaller issue blocking it, and I didn't realize it])

+++ CSCore/Main/HookManager/HookManager.Windows.cs
@@ -96,7 +96,7 @@ namespace Gma.UserActivityMonitor {
 		/// <remarks>http://msdn.microsoft.com/library/default.asp?url=/library/en-us/winui/winui/windowsuserinterface/windowing/hooks/hookreference/hookfunctions/setwindowshookex.asp</remarks>
 		[DllImport("user32.dll", CharSet = CharSet.Auto,
 			CallingConvention = CallingConvention.StdCall)] static extern int CallNextHookEx(
-			int idHook,
+			IntPtr idHook,
 			int nCode,
 			int wParam,
 			IntPtr lParam);
@@ -120,7 +120,7 @@ namespace Gma.UserActivityMonitor {
 		///     return value is NULL. To get extended error information, call GetLastError.</returns>
 		/// <remarks>http://msdn.microsoft.com/library/default.asp?url=/library/en-us/winui/winui/windowsuserinterface/windowing/hooks/hookreference/hookfunctions/setwindowshookex.asp</remarks>
 		[DllImport("user32.dll", CharSet = CharSet.Auto,
-			CallingConvention = CallingConvention.StdCall, SetLastError = true)] static extern int SetWindowsHookEx(
+			CallingConvention = CallingConvention.StdCall, SetLastError = true)] static extern IntPtr SetWindowsHookEx(
 			int idHook,
 			HookProc lpfn,
 			IntPtr hMod,
@@ -134,7 +134,7 @@ namespace Gma.UserActivityMonitor {
 		///     extended error information, call GetLastError.</returns>
 		/// <remarks>http://msdn.microsoft.com/library/default.asp?url=/library/en-us/winui/winui/windowsuserinterface/windowing/hooks/hookreference/hookfunctions/setwindowshookex.asp</remarks>
 		[DllImport("user32.dll", CharSet = CharSet.Auto,
-			CallingConvention = CallingConvention.StdCall, SetLastError = true)] static extern int UnhookWindowsHookEx(int idHook);
+			CallingConvention = CallingConvention.StdCall, SetLastError = true)] static extern int UnhookWindowsHookEx(IntPtr idHook);
 
 		/// <summary>The GetDoubleClickTime function retrieves the current double-click time for the mouse. A double-click is a
 		///     series of two clicks of the mouse button, the second occurring within a specified time after the first. The
